Separate utility methods from concat module
